refractor to try to use full static page isr
