+++ frontend/src/app/read/[wtName]/[chNum]/page.tsx
@@ -15,6 +15,8 @@ import { dbConnect } from "@/app/_utils/db";
 import DisqusComments from "@/app/_components/comments/disqus";
 import { unstable_noStore } from "next/cache";
 import ServerError from "@/app/_components/serverError";
+import ChSelectDynamicWrapper from "@/app/_components/select/selectdWrap";
+import BotChSelectDynamicWrap from "@/app/_components/select/botChSelectWrap";
 
 export async function generateStaticParams({
     params: { wtName }
@@ -110,56 +112,28 @@ async function getChContent (params:any) {
     }
 }
 
-async function getChList (params:any) {
-    unstable_noStore()
-    const wt = await Wt.findOne({slug: params.wtName })
-    if (!wt) {
-        console.log('wt not found in get content')
-        return null
-    }
-    const allWtc = await Wtc.find({wtRef: wt._id}).sort({
-        chapterNumber: -1
-    })
-    if (!allWtc) {
-        console.log('allWtc not found in get content')
-        return null
-    }
 
-    const json = {
-        chList:allWtc
-    }
-    return JSON.parse(JSON.stringify(json))
-}
 
 
 export default async function Page({params}:{params: {wtName: string; chNum: string}}) {
-    let content, chList
-    /* console.log('PARAMS FROM PG', params) */
-    const pageData = getChContent(params)
-    const chListData = getChList(params)
+    let content
+    
     try {
-        [content, chList] = await Promise.all([pageData, chListData])
+        
+        content = await getChContent(params)
     } catch (err) {
         console.error('Database Error fetching page')
+        console.error(err)
         return <ServerError></ServerError>
     }
-    /* const [content, chList] = await Promise.all([pageData, chListData]) */
-    /* const content = await getChContent(params) */
+    
     if (!content) {
         notFound()
     }
     
-    /* console.log('content:', content) */
+    
 
-    const getPrev = (num:string) => {
-        const prevNum = Number(num) - 1
-        return prevNum.toString()
-    }
-
-    const getNext =(num:string) => {
-        const nextNum = Number(num) + 1
-        return nextNum.toString()
-    }
+    
 
 
 
@@ -174,43 +148,13 @@ export default async function Page({params}:{params: {wtName: string; chNum: str
             
             
             <BreadCrumbs wtUrl={params.wtName} wtcUrl={params.chNum.toString()}></BreadCrumbs>
+            
+            <ChSelectDynamicWrapper params={params}></ChSelectDynamicWrapper>
+            
+            
         
             
-            <div className="flex flex-col w-full gap-4">
-            <div className="w-full flex flex-col py-2 px-4 items-end">
-            <ChSelect wtName={params.wtName} chList={chList.chList} curCh={params.chNum}></ChSelect>
-            </div>    
             
-            <div className="w-full flex justify-between items-center gap-4 pb-2 px-4 sm:justify-end">
-                <div className="sm:hidden">
-                <ThemeSwitcher></ThemeSwitcher>
-                </div>
-                <div className="flex gap-2">
-                {
-                    chList.chList[chList.chList.length - 1].chapterNumber < Number(params.chNum) ?
-                    <Button as={Link} href={`/read/${params.wtName}/${getPrev(params.chNum)}`} size="sm" className="text-default-500 font-semibold">
-                        {`< Prev`}
-                    </Button> :
-                    <Button as={Link} href="#" isDisabled size="sm" className="text-default-500 font-semibold">
-                    {`< Prev`}
-                </Button> 
-
-                    
-
-                }
-                {
-                    chList.chList[0].chapterNumber > Number(params.chNum) ?
-                    <Button as={Link} href={`/read/${params.wtName}/${getNext(params.chNum)}`} size="sm" className="text-default-500 font-semibold">
-                        {`Next >`}
-                    </Button>  :
-                    <Button as={Link} href="#" isDisabled size="sm" className="text-default-500 font-semibold">
-                    {`Next >`}
-                     </Button> 
-
-                }
-                </div>
-            </div>
-            </div>
             <div>
                 {content.images.map((node:any, idx:number) => {
                     console.log('CONTENT IMGS MAP', node)
@@ -223,37 +167,8 @@ export default async function Page({params}:{params: {wtName: string; chNum: str
                 })}
             </div>
             <BreadCrumbs wtUrl={params.wtName} wtcUrl={params.chNum.toString()}></BreadCrumbs>
-            <div className="flex flex-col w-full gap-4">
-                <div className="w-full flex flex-col py-1 px-4 items-end">
-                <ChSelect wtName={params.wtName} chList={chList.chList} curCh={params.chNum}></ChSelect>
-                </div>
-
-                <div className="w-full flex pb-2 px-4 justify-end gap-4 ng">
-                    {
-                        chList.chList[chList.chList.length - 1].chapterNumber < Number(params.chNum) ?
-                        <Button as={Link} href={`/read/${params.wtName}/${getPrev(params.chNum)}`} size="sm" className="text-default-500 font-semibold">
-                            {`< Prev`}
-                        </Button> :
-                        <Button as={Link} href="#" isDisabled size="sm" className="text-default-500 font-semibold">
-                        {`< Prev`}
-                    </Button> 
-
-                        
-
-                    }
-                    {
-                        chList.chList[0].chapterNumber > Number(params.chNum) ?
-                        <Button as={Link} href={`/read/${params.wtName}/${getNext(params.chNum)}`} size="sm" className="text-default-500 font-semibold">
-                            {`Next >`}
-                        </Button>  :
-                        <Button as={Link} href="#" isDisabled size="sm" className="text-default-500 font-semibold">
-                        {`Next >`}
-                        </Button> 
-
-                    }
-
-                </div>
-            </div>
+            <BotChSelectDynamicWrap params={params}></BotChSelectDynamicWrap>
+           
             <div className="text-default-500 text-xs p-4">
                 <h3>{`You are reading the Manhwa/Manga/Manhua - ${content?.wt?.altName ? content?.wt?.altName + ',' : ''} ${content.wtc.name}`}</h3>
             </div>
